Move Gengo code from EnglishText into services.
