+++ services/QuillLMS/app/workers/gengo/save_jobs_from_order_worker.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Gengo
+  class SaveJobsFromOrderWorker
+    include Sidekiq::Worker
+
+    sidekiq_options queue: SidekiqQueue::LOW
+
+    def perform(order_id)
+      SaveJobsFromOrder.run(order_id)
+    end
+  end
+end
