+++ services/QuillLMS/app/workers/gengo/save_translated_text_worker.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Gengo
+  class SaveTranslatedTextWorker
+    include Sidekiq::Worker
+
+    sidekiq_options queue: SidekiqQueue::LOW
+
+    def perform(job_id)
+      SaveTranslatedText.run(job_id)
+    end
+  end
+end
