+++ services/QuillLMS/lib/tasks/temporary/translate.rake
@@ -5,9 +5,7 @@ namespace :translate do
   task hints: :environment do
     hints = ConceptFeedback.limit(3)
     hints.map(&:create_translation_mappings)
-    jobs_list = hints.map(&:english_text).map(&:gengo_payload).compact
-    return unless jobs_list.present?
-
-    EnglishText.translate!(jobs_list:)
+    english_texts = hints.map(&:english_text)
+    Gengo::RequestTranslations.run(english_texts)
   end
 end
